extract findChain
